Add DivisionUtils.h
