+++ dbms/src/Functions/intDivOrZero.cpp
@@ -1,8 +1,6 @@
 #include <Functions/FunctionFactory.h>
 #include <Functions/FunctionBinaryArithmetic.h>
 
-#include "intDiv.h"
-
 
 namespace DB
 {
